Improvements, bug-fixes and small extensions made
- Fixed file header in every single file in the solution
- Renamed "OperationMethods" to "OperationMethod"
- Changed type of first operation's value from "object" to "string"
- Fixed wrong progresses in the UpdateInstaller
- Fixed a bug that prevented the UpdateInstaller from finding the right
dynamic link library of JSON.NET
- Fixed a bug that didn't handle operations right when handing them over
to the UpdateInstaller's process ("Unexpected token"-error)
- more...

There are 3 bugs resisting that will be fixed in the next commit which
concern a wrong display of the PopupDialog when the error message is
longer than a specific number of chars, a wrong display of operation
panels when deleting an operation node. and also there is a problem
remaining that concerns the progress of the UpdateInstaller when an
operation contains more than one "sub-operation" in a list.