Fix #48
There is an improper use of accessing UI controls in non-UI threads in the "InitializeData" method of the "ProjectEditDialog" class.

The original code: The "Popup.ShowPopup" method is called without "Invoke" in the "Task.Factory.StartNew method".

P.S. Although the BUG does not always occur, but I still think it is better to fix it.